requirements.txt
python-telegram-bot==21.6
yt-dlp
python-dotenv
sqlalchemy>=2.0
psycopg2-binary
aiohttp
qrcode[pil]
Pillow
razorpay>=1.4.2
flask>=3.0
bcrypt<4.0
